requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
django-allauth==64.5.0
drf-spectacular==0.27.3
celery==5.4.1
redis==5.2.3
requests==2.32.5
beautifulsoup4==4.13.5
Pillow==11.3.0
numpy==2.2.6
scikit-learn==1.7.2
opencv-python==4.12.0.88
elasticsearch==9.1.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.2
python-decouple==3.8
cryptography==44.0.1
dj-database-url==2.1.0